requirements.txt
torch
torchvision
gradio
grad-cam
Pillow
numpy
pandas